Refactor parser logic out into their own files and classes
