+++ src/Hive.Versioning.Tests/Ranges/SubrangeComponentTests.cs
@@ -16,14 +16,14 @@ namespace Hive.Versioning.Tests.Ranges
         private static VersionComparer ParseComparer(string input)
         {
             StringPart text = input;
-            Assert.True(VersionComparer.TryParse(ref text, out var comparer));
+            Assert.True(RangeParser.TryParseComparer(ref text, out var comparer));
             return comparer;
         }
 
         private static Subrange ParseSubrange(string input, bool valid = true)
         {
             StringPart text = input;
-            Assert.Equal(valid, Subrange.TryParse(ref text, true, out var range));
+            Assert.Equal(valid, RangeParser.TryParseSubrange(ref text, true, out var range));
             return range;
         }
